requirements.txt
torch>=0.4
torchvision
pytorch-ignite==0.1.2
yacs
